requirements.txt
django
djangorestframework
djangorestframework_jwt
django-rest-auth
django-allauth
